Nicolas Regnault: - modify monomial storage in FermionOnSphere* to match the one of BosonOnSphere*
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1254 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/Hamiltonian/ParticleOnSphereProjectorHamiltonian.cc
@@ -262,7 +262,7 @@ AbstractHamiltonian* ParticleOnSphereProjectorHamiltonian::Clone ()
 
 void ParticleOnSphereProjectorHamiltonian::EvaluateInteractionFactors()
 {
-  int* TmpMonomial = new int[this->MaxNBody];
+  unsigned long* TmpMonomial = new unsigned long[this->MaxNBody];
   int TmpSum = 0;
   double* IndexSymmetryFactor = new double [this->ProjectorSpace->GetHilbertSpaceDimension()];
   if (this->Particles->GetParticleStatistic() == ParticleOnSphere::FermionicStatistic)
