+++ FQHE/src/HilbertSpace/FermionOnSphereHaldaneBasis.cc
@@ -1245,8 +1245,8 @@ RealVector& FermionOnSphereHaldaneBasis::GenerateJackPolynomial(RealVector& jack
   jack[0] = 1.0;
   double InvAlpha =  2.0 * (1.0 - alpha) / alpha;
 
-  int* TmpMonomial = new int [this->NbrFermions];
-  int* TmpMonomial2 = new int [this->NbrFermions];
+  unsigned long* TmpMonomial = new unsigned long [this->NbrFermions];
+  unsigned long* TmpMonomial2 = new unsigned long [this->NbrFermions];
 
   double RhoRoot = 0.0;
   unsigned long MaxRoot = this->StateDescription[0];
@@ -1330,8 +1330,8 @@ RealVector& FermionOnSphereHaldaneBasis::GenerateSymmetrizedJackPolynomial(RealV
   jack[0] = 1.0;
   double InvAlpha =  2.0 * (1.0 - alpha) / alpha;
 
-  int* TmpMonomial = new int [this->NbrFermions];
-  int* TmpMonomial2 = new int [this->NbrFermions];
+  unsigned long* TmpMonomial = new  unsigned long[this->NbrFermions];
+  unsigned long* TmpMonomial2 = new unsigned long [this->NbrFermions];
 
   double RhoRoot = 0.0;
   unsigned long MaxRoot = this->StateDescription[0];
@@ -1426,7 +1426,7 @@ RealVector& FermionOnSphereHaldaneBasis::CheckPossibleSingularCoefficientsInJack
 {
   double InvAlpha =  2.0 * (1.0 - alpha) / alpha;
 
-  int* TmpMonomial = new int [this->NbrFermions];
+  unsigned long* TmpMonomial = new unsigned long [this->NbrFermions];
 
   double RhoRoot = 0.0;
   unsigned long MaxRoot = this->StateDescription[0];
